+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 - Added a test that makes sure all of the plugin's symbols resolve correctly by
   loading the library with `RTLD_NOW`. This test is only run on Unix-like
   platforms.
+- Added missing thread safety checks in the state tests.
 
 ### Changed